FractalTools.jl

FractalTools.Attractor — Type

struct Attractor{T, S, R}

Attractor of IFS type

Fields

ifs::Any

IFS of Attractor

alg::Any

Type of algorithm to be used to compute attractor(Options are DetAlg and RandAlg

initset::Any

Initial set of attractor

set::Any

Set of the attractor

numiter::Int64

Number of iterations

parallel::Bool

Sequential or parallel

FractalTools.DetAlg — Type

struct DetAlg

A type signifying that deterministic algorithm is used when calculating the attractor of and IFS.

FractalTools.IFS — Type

struct IFS{T1<:(AbstractArray{var"#s26",1} where var"#s26"<:FractalTools.Transformation), T2<:(AbstractArray{var"#s25",1} where var"#s25"<:Real)}

Iterated fucntion sytem (IFS)

Fields

ws::AbstractArray{var"#s26",1} where var"#s26"<:FractalTools.Transformation

Vector of transformations of IFS

probs::AbstractArray{var"#s25",1} where var"#s25"<:Real

Vector of probabilities of IFS

FractalTools.RandAlg — Type

struct RandAlg

A type signifying that random algorithm is used when calculating the attractor of and IFS.

FractalTools.Transformation — Type

struct Transformation{T1<:(AbstractArray{var"#s24",2} where var"#s24"<:Real), T2<:(AbstractArray{var"#s15",1} where var"#s15"<:Real)}

Affine transformation

Fields

A::AbstractArray{var"#s24",2} where var"#s24"<:Real

Transformation matrix

b::AbstractArray{var"#s15",1} where var"#s15"<:Real

Transformation vector

FractalTools.Fern — Method

Fern()

Constructs and IFS for a fern.

FractalTools.Sierpinski — Method

Sierpinski()

Conctructs an IFS for Sierpinski triangle.

FractalTools.Square — Method

Square()

Constructs and IFS for a sqaure.

FractalTools.Tree — Method

Tree()

Constructs and IFS for a fractal tree.

FractalTools.attractor — Method

attractor(ifs, initset; alg, kwargs...)

Computes the attractor of ifs. If alg is of type DetAlg, the deterministic algorithm is used. If alg is of type RandAlg, random algorithm is used. kwargs may include

numiter::Int : Number of iterations to used to calcuate the attractor (defaults to 10)

numtransient::Int : Number of transient iterations to used to calcuate a transient set. When the transient set is constructed, the computation of attractor is continued with distributed computation if alg is RandAlg and parallel is true. (defaults to 10)

parallel::Bool: If true, the attractor is computed using distrbuted computation. (defaults to false)

placedependent::Bool : If true, place dependent attractor is computed if α and β is given accordingly. (default to false)

α::AbstractVector : Place-dependent probility coefficient(defaults to nothing)

β::AbstractVector : Place-dependent probility coefficient. (default to nothing)

FractalTools.contfactor — Method

contfactor(w)

Returns contraction factor of w. Contraction factor is computed as the norm of w.A.

FractalTools.contfactor — Method

contfactor(ifs::IFS)

Returns the contraction factor of IFS.

FractalTools.detalg — Method

detalg(ifs, initset; numiter, parallel)

Computes the attractor of ifs with deterministic algorithm.numiter is number of iterations. (Defaults to 10). If parallel is true, attractor is computed via parallel computation.

FractalTools.dimension — Method

dimension(w)

Returns dimension of w.

FractalTools.dimension — Method

dimension(ifs::IFS)

Returns dimension of ifs.

FractalTools.getdata — Function

getdata(f, ngon::Ngon, npts::Int)

Returns interpolation data pnts. pnts is a vector of three-dimensional points pi = Point(xi, yi, zi) where xi and yi are from the dispersed points and zi = f(xi, yi).

getdata(ngon::Ngon, npts::Int)

Returns intepolation data that is insideo of ngon. npts is the number of points to be returned.

FractalTools.getpoint — Function

getpoint(ngon::Ngon, maxiter::Int=100_000)

Returns a valid that is inside of ngon.

getpoint(line::Line)

Returns a valid point that is inside line.

FractalTools.randalg — Method

randalg(ifs, initset; numiter, numtransient, parallel, placedependent, α, β)

Computes the attractor of ifs with random algorithm.numiter is number of iterations. (Defaults to 100). numtransient is the number of transient iterations. If parallel is true, attractor is computed via parallel computation. If placedependent is true, the probabilties of the ifs are dependent on the coordinates x. This dependency p(x) is given via the parameters α and β where p(x) = α x + β.