FractalTools.jl

FractalTools.Attractor — Type

struct Attractor{T, S, R}

Attractor of IFS type

Fields

ifs::Any

IFS of Attractor

alg::Any

Type of algorithm to be used to compute attractor(Options are DetAlg and RandAlg

initset::Any

Initial set of attractor

set::Any

Set of the attractor

numiter::Int64

Number of iterations

parallel::Bool

Sequential or parallel

FractalTools.DetAlg — Type

struct DetAlg

A type signifying that deterministic algorithm is used when calculating the attractor of and IFS.

FractalTools.HInterp1D — Type

struct HInterp1D{T<:(Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:(AbstractArray{T,2} where T)) where var"#s48"<:(AbstractArray{T,2} where T))} <: FractalTools.AbstractCurveInterp

One dimensional hidden fractal interpolation. IFS is defined as

\[ w_n(x, y) = \begin{bmatrix} a_{11, n} & 0 & 0 \ a_{21, n} & a_{22, n} & a_{23, n} \ a_{31, n} & a_{32, n} & a_{33, n} \ \end{bmatrix} \begin{bmatrix} b_{1, n} \ b_{2, n} \ b_{3, n} \ \end{bmatrix} \quad n = 1, 2, \ldots, N \]

Here

\[ \begin{bmatrix} a_{22, n} & a_{23, n} \ a_{32, n} & a_{33, n} \end{bmatrix} \]

are the free variables. While constructing an HInterp1D, the free variables can be specfied as either a single real matrix or a vector of real matrices. If a single real matrix is specified, it assumed that

\[ \begin{bmatrix} a_{22} & a_{23} \ a_{32} & a_{33} \end{bmatrix} \begin{bmatrix} a_{22, n} & a_{23, n} \ a_{32, n} & a_{33, n} \end{bmatrix} \quad \forall n = 1, 2, \ldots, N\]

Fields

freevars::Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:(AbstractArray{T,2} where T)) where var"#s48"<:(AbstractArray{T,2} where T)

Free variables

FractalTools.HInterp2D — Type

struct HInterp2D{T<:(Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:(AbstractArray{T,2} where T)) where var"#s48"<:(AbstractArray{T,2} where T))} <: FractalTools.AbstractSurfaceInterp

Two dimensional hidden fractal interpolation. IFS is defined as

\[ w_n(x, y) = \begin{bmatrix} a_{11, n} & 0 & 0 & 0 \ a_{21, n} & a_{22, n} & 0 & 0 \ a_{31, n} & a_{32, n} & a_{33, n} & a_{34, n} \ a_{41, n} & a_{42, n} & a_{43, n} & a_{44, n} \ \end{bmatrix} \begin{bmatrix} b_{1, n} \ b_{2, n} \ b_{3, n} \ b_{4, n} \ \end{bmatrix} \quad n = 1, 2, \ldots, N \]

Here

\[ \begin{bmatrix} a_{33, n} & a_{34, n} \ a_{43, n} & a_{44, n} \end{bmatrix} \]

are the free variables. While constructing an HInterp2D, the free variables can be specfied as either a single real matrix or a vector of real matrices. If a single real matrix is specified, it assumed that

\[ \begin{bmatrix} a_{33} & a_{34} \ a_{43} & a_{44} \end{bmatrix} \begin{bmatrix} a_{33, n} & a_{34, n} \ a_{43, n} & a_{44, n} \end{bmatrix} \quad \forall n = 1, 2, \ldots, N\]

Fields

freevars::Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:(AbstractArray{T,2} where T)) where var"#s48"<:(AbstractArray{T,2} where T)

Free variables

FractalTools.IFS — Type

struct IFS{T1<:(AbstractArray{var"#s26",1} where var"#s26"<:FractalTools.Transformation), T2<:(AbstractArray{var"#s25",1} where var"#s25"<:Real)}

Iterated fucntion sytem (IFS)

Fields

ws::AbstractArray{var"#s26",1} where var"#s26"<:FractalTools.Transformation

Vector of transformations of IFS

probs::AbstractArray{var"#s25",1} where var"#s25"<:Real

Vector of probabilities of IFS

FractalTools.Interp1D — Type

struct Interp1D{T<:(Union{var"#s26", var"#s25"} where var"#s25"<:(AbstractArray{var"#s24",1} where var"#s24"<:Real) where var"#s26"<:Real)} <: FractalTools.AbstractCurveInterp

One dimensional fractal interpolation. IFS is defined as

\[ w_n(x, y) = \begin{bmatrix} a_{11, n} & 0 \ a_{21, n} & a_{22, n} \end{bmatrix} \begin{bmatrix} b_{1, n} \ b_{2, n} \end{bmatrix} \quad n = 1, 2, \ldots, N \]

Here $a_{22, n}, n = 1, 2, \ldots, N$ are the free variables. While constructing an Interp1D, the free varaibles can be specfied as either a single real value or a vector of real values. If a single real value is specified, it assumed that $a_{22} = a_{22, n}, n = 1, 2, \ldots, N$.

Fields

freevars::Union{var"#s26", var"#s25"} where var"#s25"<:(AbstractArray{var"#s24",1} where var"#s24"<:Real) where var"#s26"<:Real

Free variables

FractalTools.Interp2D — Type

struct Interp2D{T<:(Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:Real) where var"#s48"<:Real)} <: FractalTools.AbstractSurfaceInterp

Two dimensional fractal interpolation. IFS is defined as

\[ w_n(x, y) = \begin{bmatrix} a_{11, n} & a_{12, n} & 0 \ a_{21, n} & a_{22, n} & 0 \ a_{31, n} & a_{32, n} & a_{33, n} \ \end{bmatrix} \begin{bmatrix} b_{1, n} \ b_{2, n} \ b_{3, n} \ \end{bmatrix} \quad n = 1, 2, \ldots, N \]

Here $a_{33,n}, n = 1, 2, \ldots, N` are the free variables. While constructing an `InterpdD`, the free variables can be specfied as either a single real value or a vector of real values. If a single real value is specified, it assumed that$a_{33, n}, n = 1, 2, \ldots, N``.

Fields

freevars::Union{var"#s48", var"#s47"} where var"#s47"<:(AbstractArray{var"#s46",1} where var"#s46"<:Real) where var"#s48"<:Real

Free variables

FractalTools.Interpolant — Type

struct Interpolant{T1<:IFS, T2, T3<:FractalTools.AbstractInterp} <: Function

Interpolation object.

Fields

ifs::IFS

IFS

itp::Any

Interpolation function

method::FractalTools.AbstractInterp

Interpolation method

FractalTools.RandAlg — Type

struct RandAlg

A type signifying that random algorithm is used when calculating the attractor of and IFS.

FractalTools.Transformation — Type

struct Transformation{T1<:(AbstractArray{var"#s24",2} where var"#s24"<:Real), T2<:(AbstractArray{var"#s15",1} where var"#s15"<:Real)}

Affine transformation

Fields

A::AbstractArray{var"#s24",2} where var"#s24"<:Real

Transformation matrix

b::AbstractArray{var"#s15",1} where var"#s15"<:Real

Transformation vector

FractalTools.Fern — Method

Fern()

Constructs and IFS for a fern.

FractalTools.Sierpinski — Method

Sierpinski()

Conctructs an IFS for Sierpinski triangle.

FractalTools.Square — Method

Square()

Constructs and IFS for a sqaure.

FractalTools.Tree — Method

Tree()

Constructs and IFS for a fractal tree.

FractalTools.attractor — Method

attractor(ifs, initset; alg, kwargs...)

Computes the attractor of ifs. If alg is of type DetAlg, the deterministic algorithm is used. If alg is of type RandAlg, random algorithm is used. kwargs may include

numiter::Int : Number of iterations to used to calcuate the attractor (defaults to 10)

numtransient::Int : Number of transient iterations to used to calcuate a transient set. When the transient set is constructed, the computation of attractor is continued with distributed computation if alg is RandAlg and parallel is true. (defaults to 10)

parallel::Bool: If true, the attractor is computed using distrbuted computation. (defaults to false)

placedependent::Bool : If true, place dependent attractor is computed if α and β is given accordingly. (default to false)

α::AbstractVector : Place-dependent probility coefficient(defaults to nothing)

β::AbstractVector : Place-dependent probility coefficient. (default to nothing)

FractalTools.contfactor — Method

contfactor(w)

Returns contraction factor of w. Contraction factor is computed as the norm of w.A.

FractalTools.contfactor — Method

contfactor(ifs)

Returns the contraction factor of IFS.

FractalTools.detalg — Method

detalg(ifs, initset; numiter, parallel)

Computes the attractor of ifs with deterministic algorithm.numiter is number of iterations. (Defaults to 10). If parallel is true, attractor is computed via parallel computation.

FractalTools.dimension — Method

dimension(w)

Returns dimension of w.

FractalTools.dimension — Method

dimension(ifs)

Returns dimension of ifs.

FractalTools.getdata — Method

getdata(f, vtx, npts)

Returns a vector of (d + m)-dimensional points that are dispersed randomly over d-dimensioanl domain with vertex points vtx. f is the vector-valued function used to map each point in the domain. npts is the number of points to returned.

getdata(vtx, npts)

Returns a vector of d-dimensional points that are dispersed randomly over d-dimensioanl domain with vertex points vtx. npts is the number of points to be returned`

FractalTools.getpoint — Function

getpoint(ngon::Ngon, maxiter::Int=100_000)

Returns a valid that is inside of ngon.

getpoint(line::Line)

Returns a valid point that is inside line.

FractalTools.integrate — Method

integrate(pts, method)

Calculates numerical integration of pts using the fractal integration method method that may be

Interp1D

HInterp1D

Interp2D

HInterp2D

FractalTools.interpolate — Method

interpolate(pts, method; f0, niter)

Returns an Interpolant by interpolating pts using method. f0 is the initial function. niter is the number of iterations during the construction of the interpolant. method can be

Interp1D

HInterp1D

Interp2D

HInterp2D

Example

julia> xi, dx, xf = 0., 0.01, 10.;

julia> x = collect(xi : dx : xf);

julia> y = sin.(x);

julia> pts = collect.(zip(x, y)); 

julia> interp = integrate(pts, Interp1D(0.001));  # One dimensional interpolation. 

FractalTools.project — Function

project(pts)
project(pts, drop)

Projects pts on a lower dimensioanl space by dropping the drop number of indexes.

FractalTools.randalg — Method

randalg(ifs, initset; numiter, numtransient, parallel, placedependent, α, β)

Computes the attractor of ifs with random algorithm.numiter is number of iterations. (Defaults to 100). numtransient is the number of transient iterations. If parallel is true, attractor is computed via parallel computation. If placedependent is true, the probabilties of the ifs are dependent on the coordinates x. This dependency p(x) is given via the parameters α and β where p(x) = α x + β.